pyproject.toml
[tool.poetry]
name = "ds-webapp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.1.0"
flask-restful = "^0.3.10"
flasgger = "^0.9.7.1"
pytest = "^8.3.5"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
responses = "^0.25.7"
pydantic = "^2.10.6"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pylint = "^3.3.5"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["ds_webapp/tests"]

[tool.pylint."MESSAGES CONTROL"]
disable = ["W0621"]
